+++ extensions/quick-access/src/pinned-folders-menu-bar.tsx
@@ -1,37 +1,11 @@
-import {
-  Action,
-  ActionPanel,
-  getPreferenceValues,
-  Icon,
-  List,
-  LocalStorage,
-  MenuBarExtra,
-  open,
-  openCommandPreferences,
-  showHUD,
-  showInFinder,
-  showToast,
-  Toast,
-} from "@raycast/api";
+import { getPreferenceValues, Icon, MenuBarExtra, open, openCommandPreferences, showHUD } from "@raycast/api";
 import React, { useState } from "react";
-import { getLocalStorage, isEmpty, isImage } from "./utils/common-utils";
-import { DirectoryInfo, DirectoryWithFileInfo, FileInfo, FileType } from "./types/types";
+import { isImage } from "./utils/common-utils";
+import { DirectoryWithFileInfo } from "./types/types";
 import { parse } from "path";
 import { pinFolder } from "./pin-folder";
-import { LocalStorageKey, tagDirectoryTypes } from "./utils/constants";
-import {
-  alertDialog,
-  copyLatestFile,
-  getFileInfoAndPreview,
-  getIsShowDetail,
-  localDirectoryWithFiles,
-  refreshNumber,
-} from "./hooks/hooks";
-import { ActionRemoveAllDirectories, ActionsOnFile } from "./components/action-on-files";
+import { localDirectoryWithFiles, refreshNumber } from "./hooks/hooks";
 import { Preferences } from "./types/preferences";
-import { ActionOpenCommandPreferences } from "./components/action-open-command-preferences";
-import { QuickAccessEmptyView } from "./components/quick-access-empty-view";
-import { ItemDetail } from "./components/item-detail";
 import { copyFileByPath } from "./utils/applescript-utils";
 
 export default function SearchPinnedFolders() {
@@ -96,19 +70,24 @@ function FolderMenuBarItem(props: { directory: DirectoryWithFileInfo; primaryAct
           icon={isImage(parse(fileValue.path).ext) ? { source: fileValue.path } : { fileIcon: fileValue.path }}
           title={fileValue.name}
           tooltip={fileValue.path}
-          onAction={async () => {
+          onAction={async (event: MenuBarExtra.ActionEvent) => {
             switch (primaryAction) {
               case "Copy": {
-                await showHUD(`${fileValue.name} is copied to clipboard`);
-                await copyFileByPath(fileValue.path);
-                break;
-              }
-              case "Show": {
-                await showInFinder(fileValue.path);
+                if (event.type == "left-click") {
+                  await showHUD(`${fileValue.name} is copied to clipboard`);
+                  await copyFileByPath(fileValue.path);
+                } else {
+                  await open(fileValue.path);
+                }
                 break;
               }
               case "Open": {
-                await open(fileValue.path);
+                if (event.type == "left-click") {
+                  await open(fileValue.path);
+                } else {
+                  await showHUD(`${fileValue.name} is copied to clipboard`);
+                  await copyFileByPath(fileValue.path);
+                }
                 break;
               }
             }