+++ extensions/todoist/src/components/View.tsx
@@ -1,0 +1,10 @@
+import { useEffect } from "react";
+import { checkTodoistApp } from "../helpers/isTodoistInstalled";
+
+export default function View({ children }: { children: JSX.Element }) {
+  useEffect(() => {
+    checkTodoistApp();
+  }, []);
+
+  return children;
+}